# Topology spreading scoring with automatically weighted topologies